Add cidc_api.config package to setup.py (#25)
