+++ cidc_api/app.py
@@ -13,7 +13,7 @@ from auth import BearerAuth
 from services import register_services
 
 ABSPATH = dirname(abspath(__file__))
-SETTINGS = join(ABSPATH, "settings.py")
+SETTINGS = join(ABSPATH, "config", "settings.py")
 MIGRATIONS = join(ABSPATH, "..", "migrations")
 
 # Instantiate the Eve app
