+++ setup.py
@@ -3,11 +3,16 @@
 from setuptools import setup
 
 setup(
-    name="cidc_api_models",
+    name="cidc_api_modules",
     description="SQLAlchemy data models used in the CIDC API",
     python_requires=">=3.6",
-    install_requires=["flask-sqlalchemy==2.4.0", "flask==1.1.1"],
+    install_requires=[
+        "flask==1.1.1",
+        "flask-sqlalchemy==2.4.0",
+        "google-cloud-storage==1.16.1",
+    ],
     license="MIT license",
+    packages=["cidc_api.config"],
     py_modules=["cidc_api.models"],
     url="https://github.com/CIMAC-CIDC/cidc_api-gae",
     version="0.1.0",
